requirement.txt
Django==4.2.2
djangorestframework==3.14.0
transformers==4.30.2
torch==2.0.1